Renamed package constants to constant (singular)
