+++ tests/unit/test_ForeverOrderEndpoint.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from dhanhq.constants import ExchangeSegment, LegName, OrderFlag, OrderType, ProductType, TransactionType, Validity
+from dhanhq.constant import ExchangeSegment, LegName, OrderFlag, OrderType, ProductType, TransactionType, Validity
 
 
 class TestForeverOrderEndpoint:
