+++ tests/unit/test_FundsEndpoint.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from dhanhq.constants import ExchangeSegment, ProductType, TransactionType
+from dhanhq.constant import ExchangeSegment, ProductType, TransactionType
 
 
 class TestFundsEndpoint:
