+++ tests/unit/test_OrderEndpoint.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from dhanhq.constants import AMOTime, ExchangeSegment, LegName, OrderType, ProductType, TransactionType, Validity
+from dhanhq.constant import AMOTime, ExchangeSegment, LegName, OrderType, ProductType, TransactionType, Validity
 
 
 class TestOrderEndpoint:
