+++ tests/unit/test_PortfolioEndpoint.py
@@ -1,6 +1,6 @@
 from unittest.mock import patch
 
-from dhanhq.constants import ExchangeSegment, PositionType, ProductType
+from dhanhq.constant import ExchangeSegment, PositionType, ProductType
 
 
 class TestPortfolioEndpoint:
